fix(dvh/types): fix canonical_key collision, Python 3.10 timestamp, mypy and pylint errors
BUG-1: ProvenanceRecord.timestamp_utc now normalises 'Z' → '+00:00'
before calling datetime.fromisoformat(), fixing a runtime ValueError
on Python 3.10 where 'Z' is not accepted.

BUG-2: MetricSpec.canonical_key now includes index_metric when
family=INDEX, preventing HI/CI/PCI/GI from colliding and allowing
multiple index metrics per ROI (a standard clinical workflow).

BUG-3: Remove unused imports _validate_positive_finite and timezone
from _results.py (flagged by pylint W0611).

DESIGN-1: MetricRequestSet.to_dict() now delegates to
DoseReferenceSet.to_dict() for consistent serialisation schema;
from_dict() accepts both new and legacy formats.

DESIGN-2: MetricSpec(family=INDEX) now raises ValueError if
index_metric cannot be inferred from raw, preventing unknown
index metrics from silently getting requires_dose_ref=False.

DESIGN-3: DoseReferenceSet validates keys before freezing with
MappingProxyType, not after.

DESIGN-4: ROIResult.to_dict() simplified — redundant isinstance
guard removed since __post_init__ always coerces to ROIStatus.

DESIGN-5: ROIDiagnostics rejects bool values for integer count
fields (bool is a subclass of int in Python).

DESIGN-6: DVHBins.__hash__ removes redundant float() on
total_volume_cc.

MYPY-1: GridFrame.from_dict() returns a 3-tuple via indexing
instead of tuple(list), fixing mypy arg-type error.

MYPY-2: Geometry volume functions wrap numpy returns in float()
to satisfy mypy no-any-return checks.

406 tests pass (11 new regression tests added).

+++ lib/pymedphys/_dvh/_benchmarks/_geometry.py
@@ -60,7 +60,7 @@ def sphere_volume(radius_mm: float) -> float:
         Volume in mm³: ``(4/3) * π * r³``.
     """
     _validate_positive(radius_mm=radius_mm)
-    return (4.0 / 3.0) * np.pi * radius_mm**3
+    return float((4.0 / 3.0) * np.pi * radius_mm**3)
 
 
 def cylinder_volume(radius_mm: float, height_mm: float) -> float:
@@ -79,7 +79,7 @@ def cylinder_volume(radius_mm: float, height_mm: float) -> float:
         Volume in mm³: ``π * r² * h``.
     """
     _validate_positive(radius_mm=radius_mm, height_mm=height_mm)
-    return np.pi * radius_mm**2 * height_mm
+    return float(np.pi * radius_mm**2 * height_mm)
 
 
 def cone_volume(radius_mm: float, height_mm: float) -> float:
@@ -98,7 +98,7 @@ def cone_volume(radius_mm: float, height_mm: float) -> float:
         Volume in mm³: ``(1/3) * π * r² * h``.
     """
     _validate_positive(radius_mm=radius_mm, height_mm=height_mm)
-    return (1.0 / 3.0) * np.pi * radius_mm**2 * height_mm
+    return float((1.0 / 3.0) * np.pi * radius_mm**2 * height_mm)
 
 
 def ellipsoid_volume(semi_a_mm: float, semi_b_mm: float, semi_c_mm: float) -> float:
@@ -119,7 +119,7 @@ def ellipsoid_volume(semi_a_mm: float, semi_b_mm: float, semi_c_mm: float) -> fl
         Volume in mm³: ``(4/3) * π * a * b * c``.
     """
     _validate_positive(semi_a_mm=semi_a_mm, semi_b_mm=semi_b_mm, semi_c_mm=semi_c_mm)
-    return (4.0 / 3.0) * np.pi * semi_a_mm * semi_b_mm * semi_c_mm
+    return float((4.0 / 3.0) * np.pi * semi_a_mm * semi_b_mm * semi_c_mm)
 
 
 def torus_volume(major_radius_mm: float, minor_radius_mm: float) -> float:
@@ -156,7 +156,7 @@ def torus_volume(major_radius_mm: float, minor_radius_mm: float) -> float:
             f"than minor_radius_mm ({minor_radius_mm}) for a simple "
             f"(non-self-intersecting) torus"
         )
-    return 2.0 * np.pi**2 * major_radius_mm * minor_radius_mm**2
+    return float(2.0 * np.pi**2 * major_radius_mm * minor_radius_mm**2)
 
 
 def cylindrical_shell_volume(
@@ -193,7 +193,7 @@ def cylindrical_shell_volume(
             f"outer_radius_mm ({outer_radius_mm}) must be strictly greater "
             f"than inner_radius_mm ({inner_radius_mm})"
         )
-    return np.pi * (outer_radius_mm**2 - inner_radius_mm**2) * height_mm
+    return float(np.pi * (outer_radius_mm**2 - inner_radius_mm**2) * height_mm)
 
 
 def rectangular_parallelepiped_volume(
